fixed typo

+++ engine/code/src/gfx/vkrenderer.cpp
@@ -824,7 +824,7 @@ void Gfx::Renderer::PrepareCameraBuffer(Keeper::Camera& camera)
             float linear = 0.7;
             float quadratic = 1.8;
             float light_max = std::fmaxf(std::fmaxf(obj->GetColor().x, obj->GetColor().y), obj->GetColor().z);
-            float radius = (-linear + std::sqrtf(linear * linear - 4 * quadratic * (constant - (256.0 / 5.0) * light_max))) / (2 * quadratic);
+            float radius = (-linear + std::sqrt(linear * linear - 4 * quadratic * (constant - (256.0 / 5.0) * light_max))) / (2 * quadratic);
             if (j >= 4) {
                 j = 0;
             }
